+++ docs/history.md
@@ -18,6 +18,7 @@ V1.0.8
 identification, authentication, and encryption (IKEv1 or IKEv2, Phase 1).
 -   panos\_ipsec\_profile: Configures IPSec Crypto profile on the firewall with subset of settings.
 -   panos\_ike\_gateway: Configures IKE gateway on the firewall with subset of settings.
+-   panos\_ipsec\_tunnel: Configure data-port (DP) network interface for DHCP. By default DP interfaces are static.
 
 *Removed modules*